Added pagination in all books, genre display, new arrivals, best sellers. Added votes number in reviews. Added animation when new announcment arrives. Added confirmation when removing books from shelves. Prettified some files.

+++ frontend/scripts/controllers/book.js
@@ -1,33 +1,33 @@
-
-
 'use strict';
 
 ///Book Controller's main job is to recieve a book and prepare it to be displayed with all its info
 angular.module('angularPassportApp')
-  .controller('BookCtrl', function ($scope, $http, $location, ShareService, EpubService) {
+  .controller('BookCtrl', function($scope, $http, $location, ShareService, EpubService) {
     $scope.review = '';
-    $scope.reviews = [];    
+    $scope.reviews = [];
     $scope.tags;
 
-    $http.get('api/books/'+ShareService.getValue()).success(function(response) {
+    $http.get('api/books/' + ShareService.getValue()).success(function(response) {
       console.log("I received the book");
-      $scope.book=response.book;
+      $scope.book = response.book;
       $scope.bookID = response.book._id;
-      $scope.book.tobeRead=false;
+      $scope.book.tobeRead = false;
       console.log(response.book.labels);
       $scope.tags = response.book.labels;
 
       //function to get the avg rating of a book
-      $http.get('api/books/'+ShareService.getValue()+'/avgRating').success(function(rating) {
-      console.log("avg rating is " + rating);
-      $scope.book.avgRating=rating;
+      $http.get('api/books/' + ShareService.getValue() + '/avgRating').success(function(rating) {
+        console.log("avg rating is " + rating);
+        $scope.book.avgRating = rating;
       });
 
       //get the rating of the current user of that book
-      $http.post('api/books/'+ShareService.getValue()+'/getrate', {userId: $scope.currentUser._id})
+      $http.post('api/books/' + ShareService.getValue() + '/getrate', {
+          userId: $scope.currentUser._id
+        })
         .success(function(num) {
           $scope.rating = {}
-          $scope.rating=num;
+          $scope.rating = num;
 
           console.log("user rating is " + num);
         })
@@ -37,35 +37,37 @@ angular.module('angularPassportApp')
 
 
       /**
-      * @function getIsTobeRead
-      * @determines whether the book is on the current user tobe read list or not
-      * @param {boolean} bool - is Book on to-be read list
-      */
-      $http.post('api/books/'+ShareService.getValue()+'/istoberead', {userId: $scope.currentUser._id})
+       * @function getIsTobeRead
+       * @determines whether the book is on the current user tobe read list or not
+       * @param {boolean} bool - is Book on to-be read list
+       */
+      $http.post('api/books/' + ShareService.getValue() + '/istoberead', {
+          userId: $scope.currentUser._id
+        })
         .success(function getIsTobeRead(bool) {
-          $scope.book.tobeRead=bool;
+          $scope.book.tobeRead = bool;
           console.log("user added book " + bool);
         })
         .error(function(data) {
           console.log('Error: ' + data);
         });
 
-        $http.get('api/books/' + ShareService.getValue() + '/review')
-      .success(function (response) {
-        $scope.reviews = response;
-        console.log('$scope.reviews', $scope.reviews);
-      }).error(function(data) {
+      $http.get('api/books/' + ShareService.getValue() + '/review')
+        .success(function(response) {
+          $scope.reviews = response;
+          console.log('$scope.reviews', $scope.reviews);
+        }).error(function(data) {
           console.log('Error: ' + data);
         });
     });
 
-    
 
-  /**
-   * @function rate  
-   * sends user rating to backend
-   * @param {int} userRating - user rating
-   */
+
+    /**
+     * @function rate  
+     * sends user rating to backend
+     * @param {int} userRating - user rating
+     */
     $scope.rate = function(userRating) {
       console.log('user: ' + $scope.currentUser._id);
       console.log('book: ' + ShareService.getValue());
@@ -88,40 +90,56 @@ angular.module('angularPassportApp')
 
 
     /**
-      * @function aadTobeRead
-      * @adds a book to user's to be read list
-    */
+     * @function aadTobeRead
+     * @adds a book to user's to be read list
+     */
     $scope.addTobeRead = function() {
-      $http.post('api/books/'+ShareService.getValue()+'/ToBeRead')
+      $http.post('api/books/' + ShareService.getValue() + '/ToBeRead')
         .success(function(response) {
-          $http.post('api/books/'+ShareService.getValue()+'/istoberead', {userId: $scope.currentUser._id})
-          .success(function(bool) {
+          $http.post('api/books/' + ShareService.getValue() + '/istoberead', {
+              userId: $scope.currentUser._id
+            })
+            .success(function(bool) {
 
-          $scope.book.tobeRead=bool;
-        })
+              $scope.book.tobeRead = bool;
+            })
 
         })
         .error(function(data) {
           console.log('Error: ' + data);
         });
 
-      };
+    };
     /**
-      * @function removeTobeRead
-      * @deletes a book from user's to be read list
-    */
+     * @function removeTobeRead
+     * @deletes a book from user's to be read list
+     */
     $scope.removeTobeRead = function() {
-      $http.delete('api/books/'+ShareService.getValue()+'/ToBeRead')
-        .success(function(response) {
-          $http.post('api/books/'+ShareService.getValue()+'/istoberead', {userId: $scope.currentUser._id})
-          .success(function(bool) {
-          $scope.book.tobeRead=bool;
-           console.log("user added book" + bool);
-            })
-        })
-        .error(function (data) {
-          console.log('Error: ' + data);
-        });
+
+      swal({
+        title: "Are you sure?",
+        type: "error",
+        showCancelButton: true,
+        confirmButtonColor: "#DD6B55",
+        confirmButtonText: "Yes, delete it!",
+        closeOnConfirm: true
+      }, function() {
+
+        $http.delete('api/books/' + ShareService.getValue() + '/ToBeRead')
+          .success(function(response) {
+            $http.post('api/books/' + ShareService.getValue() + '/istoberead', {
+                userId: $scope.currentUser._id
+              })
+              .success(function(bool) {
+                $scope.book.tobeRead = bool;
+                console.log("user added book" + bool);
+              })
+          })
+          .error(function(data) {
+            console.log('Error: ' + data);
+          });
+
+      })
     };
 
 
@@ -132,28 +150,34 @@ angular.module('angularPassportApp')
     };
 
     //writeReview sends the review along with the user and the book associated with that review
-    $scope.writeReview = function (){
-      console.log('$scope.review:' , $scope.review);
-      $http.post('api/books/' + $scope.book._id + '/review',{'review' : $scope.review })
-      .success(function (response) {
-        $scope.reviews.push(response);
-        console.log('$scope.reviews', $scope.reviews);
-      })
+    $scope.writeReview = function() {
+      console.log('$scope.review:', $scope.review);
+      $http.post('api/books/' + $scope.book._id + '/review', {
+          'review': $scope.review
+        })
+        .success(function(response) {
+          $scope.reviews.push(response);
+          console.log('$scope.reviews', $scope.reviews);
+        })
       $scope.review = '';
     };
     /*deleteReview 
     checks the index of the review to be deleted and removes it from the list of reviews.
     sends a boolean delete to be checked in the backend whether the review is to be deleted. It also sends the user id of the author of the review as well as the review id int he body.
     */
-    $scope.deleteReview = function(rev){
-      console.log('reviewId' , rev._id);
-      console.log('reviewnafso' , rev.review);
+    $scope.deleteReview = function(rev) {
+      console.log('reviewId', rev._id);
+      console.log('reviewnafso', rev.review);
 
-      $http.post('api/books/' + $scope.book._id + '/review',{delete: true, 'reviewId' : rev._id, 'userId': rev.userId._id })
-      .success(function (response) {
-        console.log('userId', rev.userId._id );
-        $scope.reviews.splice($scope.reviews.indexOf(rev),1);
-      })
+      $http.post('api/books/' + $scope.book._id + '/review', {
+          delete: true,
+          'reviewId': rev._id,
+          'userId': rev.userId._id
+        })
+        .success(function(response) {
+          console.log('userId', rev.userId._id);
+          $scope.reviews.splice($scope.reviews.indexOf(rev), 1);
+        })
     };
     /*voteReview 
     checks that user havent voted neither up nor down on the review yet.
@@ -163,38 +187,44 @@ angular.module('angularPassportApp')
       -the review id
       -both arrays of up and down votes
     */
-    $scope.voteReview = function(rev, vote){
-       if ((rev.upVotes.indexOf($scope.currentUser._id)!=-1 )||(rev.downVotes.indexOf($scope.currentUser._id)!=-1)){
+    $scope.voteReview = function(rev, vote) {
+      console.log('REVIEW', rev);
+      if ((rev.upVotes.indexOf($scope.currentUser._id) != -1) || (rev.downVotes.indexOf($scope.currentUser._id) != -1)) {} else {
+        if (vote == "up") {
+          rev.upVotes.push(rev.userId._id);
+        } else {
+          rev.downVotes.push(rev.userId._id);
+        }
+      }
+      $http.post('api/books/' + $scope.book._id + '/review', {
+          vote: true,
+          'reviewId': rev._id,
+          'userId': $scope.currentUser._id,
+          'upV': rev.upVotes,
+          'downV': rev.downVotes
+        })
+        .success(function(response) {
+          console.log('userId', rev.userId._id);
+          /*var currentRev = $scope.reviews.find(rev);*/
 
-        }
-        else{ if (vote == "up"){
-              rev.upVotes.push(rev.userId._id);          
-              }
-              else{
-                rev.downVotes.push(rev.userId._id);
-              }
-        }
-      $http.post('api/books/' + $scope.book._id + '/review',{vote: true, 'reviewId' : rev._id, 'userId': $scope.currentUser._id, 'upV': rev.upVotes, 'downV': rev.downVotes})
-      .success(function (response) {
-        console.log('userId', rev.userId._id );
-        /*var currentRev = $scope.reviews.find(rev);*/
-       
-      });
+        });
     };
 
-     /**
-  * @function editTags Called on tag-added or tag-removed
-  * Changes labels after a label is added or removed
-  */
-      $scope.editTags = function() {
-       var ss = $scope.tags;
-       console.log(ss);
-        $http.post('/api/books/' + $scope.bookID + '/labels', {labels: ss}).success(function(response){
+    /**
+     * @function editTags Called on tag-added or tag-removed
+     * Changes labels after a label is added or removed
+     */
+    $scope.editTags = function() {
+      var ss = $scope.tags;
+      console.log(ss);
+      $http.post('/api/books/' + $scope.bookID + '/labels', {
+        labels: ss
+      }).success(function(response) {
         console.log(response);
       }).error(function(data) {
-          console.log('Errorlabel: ' + data);
-        });
-          };
+        console.log('Errorlabel: ' + data);
+      });
+    };
 
 
 
@@ -224,10 +254,10 @@ angular.module('angularPassportApp')
      * Saves the book in epubService
      */
     $scope.broadCastEpub = function broadCastEpub() {
-            console.log('SCOPE BOOK', $scope.book);
-            EpubService.setValue($scope.book);
-            $location.path('/bookreader');
-        }
+      console.log('SCOPE BOOK', $scope.book);
+      EpubService.setValue($scope.book);
+      $location.path('/bookreader');
+    }
